Finished login and register pages + connection to MongoDB to these pages
(Note: will now work on linking to Spotify Login again which the user should be prompted with rather than going directly to the home page - which is how the code is set to for now).

Also fixed authentication so that the navbar displays the appropriate pages depending on whether user is authenticated or not.